+++ src/backend/joanie/tests/core/test_api_organizations_contract.py
@@ -12,6 +12,8 @@ from joanie.tests.base import BaseAPITestCase
 class OrganizationContractApiTest(BaseAPITestCase):
     """Test suite for the Organizations Contract API"""
 
+    maxDiff = None
+
     def test_api_organizations_contracts_list_anonymous(self):
         """
         Anonymous user cannot query all contracts from an organization.
@@ -48,7 +50,7 @@ class OrganizationContractApiTest(BaseAPITestCase):
             )
 
         self.assertEqual(response.status_code, 200)
-        self.assertEqual(
+        self.assertDictEqual(
             response.json(),
             {
                 "count": 0,
@@ -87,7 +89,7 @@ class OrganizationContractApiTest(BaseAPITestCase):
             )
 
         self.assertEqual(response.status_code, 200)
-        self.assertEqual(
+        self.assertDictEqual(
             response.json(),
             {
                 "count": 0,
@@ -142,7 +144,7 @@ class OrganizationContractApiTest(BaseAPITestCase):
         self.assertEqual(response.status_code, 200)
         contracts = models.Contract.objects.filter(order__organization=organizations[0])
         expected_contracts = sorted(contracts, key=lambda x: x.created_on, reverse=True)
-        self.assertEqual(
+        self.assertDictEqual(
             response.json(),
             {
                 "count": 5,
@@ -179,8 +181,8 @@ class OrganizationContractApiTest(BaseAPITestCase):
                                 "logo": "_this_field_is_mocked",
                                 "title": contract.order.organization.title,
                             },
-                            "owner": contract.order.owner.username,
-                            "product": contract.order.product.title,
+                            "owner_name": contract.order.owner.username,
+                            "product_title": contract.order.product.title,
                         },
                     }
                     for contract in expected_contracts
@@ -377,7 +379,7 @@ class OrganizationContractApiTest(BaseAPITestCase):
 
         self.assertEqual(response.status_code, 200)
 
-        self.assertEqual(
+        self.assertDictEqual(
             response.json(),
             {
                 "id": str(contract.id),
@@ -405,8 +407,8 @@ class OrganizationContractApiTest(BaseAPITestCase):
                         "logo": "_this_field_is_mocked",
                         "title": contract.order.organization.title,
                     },
-                    "owner": contract.order.owner.username,
-                    "product": contract.order.product.title,
+                    "owner_name": contract.order.owner.username,
+                    "product_title": contract.order.product.title,
                 },
             },
         )